Remove old StringParser
