+++ tests/TerminationParser.php
@@ -2,7 +2,7 @@
 
 namespace webignition\StringParser\Tests;
 
-use webignition\StringParser\ConcreteStringParser;
+use webignition\StringParser\StringParser;
 
 /**
  * A demonstration parser that returns exactly what is has been given up to a chosen character limit.
@@ -11,17 +11,17 @@ class TerminationParser
 {
     private const STATE_IN_VALUE = 1;
 
-    private ConcreteStringParser $stringParser;
+    private StringParser $stringParser;
     private int $count = 0;
 
     public function __construct(
         private int $limit = 10,
     ) {
-        $this->stringParser = new ConcreteStringParser([
-            ConcreteStringParser::STATE_UNKNOWN => function (ConcreteStringParser $stringParser) {
+        $this->stringParser = new StringParser([
+            StringParser::STATE_UNKNOWN => function (StringParser $stringParser) {
                 $stringParser->setCurrentState(self::STATE_IN_VALUE);
             },
-            self::STATE_IN_VALUE => function (ConcreteStringParser $stringParser) {
+            self::STATE_IN_VALUE => function (StringParser $stringParser) {
                 $this->count++;
 
                 if ($this->count <= $this->limit) {
